app/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
Django==6.0
django-crispy-forms==2.5
django-recaptcha==4.1.0
et_xmlfile==2.0.0
idna==3.11
numpy==2.3.5
openpyxl==3.1.5
pillow==12.0.0
PyMySQL==1.1.2
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
python-decouple==3.8
gunicorn==21.2.0
psycopg[binary]>=3.1
